bootctl: Add --force option to enable chroot install/remove of sd-boot
This change enables the force-installation, removal, and update of the
systemd EFI boot binaries and supporting assets, without requiring verification
of the ESP.

Prior to this change it is assumed the target path is an ESP that can be probed,
however this requires several virtual filesystems to be present and mounted,
which may not be possible during filesystem image generation, or even for use
within eltorito alt sections of an ISO.

Given that an image invocation of bootctl may involve a host that has older,
or even absent files required by bootctl (such as EFI files), it is not always
safe to assume one can use the --path option. This change ensures that the
utility can be used in a consistent manner, in a bare chroot (no virtual
filesystems), without the need for custom systemd-boot installation routines
currently seen in many places; i.e bootctl both sets and enforces the standard
for its installation, configuration and layout.